backend/birdiepedia/requirements.txt
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==1.10.0
beautifulsoup4==4.11.1
bleach==6.0.0
blinker==1.6.3
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.6
djangorestframework==3.14.0
executing==1.2.0
ez-setup==0.9
fastjsonschema==2.18.0
filelock==3.12.4
Flask==3.0.0
fonttools==4.42.1
fqdn==1.5.1
geojson==3.0.1
google-api-core==2.11.0
google-auth==2.16.1
googleapis-common-protos==1.58.0
idna==3.4
imagesize==1.4.1
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
labelbox==3.38.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.1
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
ndjson==0.3.1
nest-asyncio==1.5.8
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.26.0
opencv-python==4.7.0.72
overrides==7.4.0
packaging==23.0
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.22.0
psutil==5.9.5
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
pyFirmata==1.1.0
pyGeoTile==1.0.6
Pygments==2.16.1
pyparsing==3.1.1
pyproj==3.4.1
pyserial==3.5
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.2
Send2Trash==1.8.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.64.1
traitlets==5.10.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.13
virtualenv==20.24.5
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.0
widgetsnbextension==4.0.9
